requirements.txt
python-dotenv==1.0.0
pinecone-client==2.2.1
openai==0.28.1
langchain==0.0.330